Implemented JGLFW on Linux and misc generation improvements:
- Added support for invoking macros as functions.
- Added support for custom (not defined in any header) struct definitions.
- Allowed different Java and native names for structs.
- Major nested structs refactoring.

+++ src/core/org/lwjgl/system/glfw/WindowCallbackAdapter.java
@@ -44,7 +44,7 @@ public abstract class WindowCallbackAdapter extends WindowCallback {
 	}
 
 	@Override
-	public void cursorPos(long window, int xpos, int ypos) {
+	public void cursorPos(long window, double xpos, double ypos) {
 	}
 
 	@Override
